# docs/swagger.yaml
basePath: /api
definitions:
  request.ApiAddReq:
    properties:
      category:
        maxLength: 50
        minLength: 1
        type: string
      method:
        maxLength: 20
        minLength: 1
        type: string
      path:
        maxLength: 100
        minLength: 1
        type: string
      remark:
        maxLength: 100
        minLength: 0
        type: string
    required:
      - category
      - method
      - path
    type: object
  request.ApiDeleteReq:
    properties:
      apiIds:
        items:
          type: integer
        type: array
    required:
      - apiIds
    type: object
  request.ApiUpdateReq:
    properties:
      category:
        maxLength: 50
        minLength: 1
        type: string
      id:
        type: integer
      method:
        maxLength: 20
        minLength: 1
        type: string
      path:
        maxLength: 100
        minLength: 1
        type: string
      remark:
        maxLength: 100
        minLength: 0
        type: string
    required:
      - id
    type: object
  request.BaseChangePwdReq:
    properties:
      code:
        type: string
      mail:
        maxLength: 100
        minLength: 0
        type: string
    required:
    - code
    - mail
    type: object
  request.BaseSendCodeReq:
    properties:
      mail:
        maxLength: 100
        minLength: 0
        type: string
    required:
    - mail
    type: object
  request.FieldRelationAddReq:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      flag:
        maxLength: 20
        minLength: 1
        type: string
    required:
      - attributes
      - flag
    type: object
  request.FieldRelationDeleteReq:
    properties:
      fieldRelationIds:
        items:
          type: integer
        type: array
    required:
      - fieldRelationIds
    type: object
  request.FieldRelationUpdateReq:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      flag:
        maxLength: 20
        minLength: 1
        type: string
      id:
        type: integer
    required:
      - attributes
      - flag
      - id
    type: object
...
  response.ResponseBody:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 127.0.0.1:8888
info:
  contact:
    email: contact@example.com
    name: Example
    url: https://example.com
  description: 基于Go+Vue实现的openLDAP后台管理项目
  termsOfService: https://example.com
  title: Go Ldap Admin
  version: "1.0"
paths:
  /api/add:
    post:
      consumes:
        - application/json
      parameters:
        - description: 新建API
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.ApiAddReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 新建API接口
      tags:
        - 接口管理
  /api/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除API
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.ApiDeleteReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 删除API接口
      tags:
        - 接口管理
  /api/list:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获取API接口列表
      tags:
        - 接口管理
  /api/tree:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获取API接口树
      tags:
        - 接口管理
  /api/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: 更新API
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.ApiUpdateReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 更新API接口
      tags:
        - 接口管理
  /base/changePwd:
    post:
      consumes:
      - application/json
      description: 使用邮箱验证码修改密码
      parameters:
      - description: 发送验证码请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BaseChangePwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 用户通过邮箱修改密码
      tags:
      - 基础管理
...
  /base/decryptpwd:
    get:
      consumes:
      - application/json
      description: 将加密后的密码解密为明文
      parameters:
      - description: 需要解密的加密密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 密码解密
      tags:
      - 基础管理
  /base/encryptpwd:
    get:
      consumes:
      - application/json
      description: 将明文密码加密
      parameters:
      - description: 需要加密的明文密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 密码加密
      tags:
      - 基础管理
...
  /base/sendcode:
    post:
      consumes:
      - application/json
      description: 向指定邮箱发送验证码
      parameters:
      - description: 发送验证码请求数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BaseSendCodeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      summary: 发送验证码
      tags:
      - 基础管理
  /fieldrelation/add:
    post:
      consumes:
        - application/json
      parameters:
        - description: 新建字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationAddReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 新建字段关系管理记录
      tags:
        - 字段关系管理
  /fieldrelation/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: 删除字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationDeleteReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 删除字段关系管理记录
      tags:
        - 字段关系管理
  /fieldrelation/list:
    get:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 获字段关系管理列表
      tags:
        - 字段关系管理
  /fieldrelation/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: 更新字段关系管理记录
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.FieldRelationUpdateReq'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBody'
      security:
        - ApiKeyAuth: [ ]
      summary: 更新字段关系管理记录
      tags:
        - 字段关系管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"